Closed #19
Issue was related to commas being in possible readings. Like: kyuu, ku .
I had assumed that each vocab instance had ONE correct answer.
Modified DB_repopulator to reflect this, and create multiple readings
for each vocab that has a comma in it. Effectively as if they were two
seperate vocab. Tidied up displaying of Hiragana/Kanji on the review
page, fixed a small bug in the unlock functionality where we werent
calling distinct() on the dataset for the count of unlocked vocabulary.