Re-unfixed crippling bug.
My previous patch caused an occasional race condition on the
UserSpecific Model, which never occured for me locally, but showed up
when lag times are slightly higher when using a remote server. Put
everything back into one ajax call, with a tracer that indicates whether
or not the user has been previously wrong.